scripts/requirements.txt
tornado==6.4.1
paho-mqtt
pytz
pymongo
motor
openpyxl
psutil
